change package name and add contact file
