+++ contacts/csv.go
@@ -1,4 +1,4 @@
-package csv
+package contacts
 
 import (
 	"encoding/csv"
@@ -10,18 +10,6 @@ import (
 	"strings"
 	"time"
 )
-
-type Contact struct {
-	FirstName     string
-	LastName      string
-	HomePhone     string
-	MobilePhone   string
-	StreetAddress string
-	City          string
-	State         string
-	Zip           string
-	BirthMonth    time.Month
-}
 
 const (
 	firstNameIdx int = iota
