+++ contacts/csv_test.go
@@ -1,4 +1,4 @@
-package csv
+package contacts
 
 import (
 	"strings"
